Create crop_image.py
This program takes files from a directory called images and then crops them. Images should be sequentially named for this code to work.  It first asks the user the number of images in the folder to crop. It then creates a matplotlib interactive environment where the user can then select the area to crop. The first click is for the top left and the second click is for the bottom right. The first image will be completely used for the crop. The subsequent images will use the same crop area but only use the left half of the crop area. It will then stitch all the images together.