Create batch_test.py
This tests a gpu's capabilities by using a simple Convolutional Neural Network and data loader. It starts with smaller batch sizes and then doubles it each time until a GPU memory error occurs. This suggests the previous batch size. Note that for real applications much lower batch sizes should be used and that whatever batch size is used the model should be watched closely so the results do not change. Sometimes your learning rate or other hyperparameters need to be changed to retain the model's robustness as an increased batch size can affect its accuracy. While Running this code it is recommended to uses a task manager and monitor GPU usage and GPU memory.