Create class_index.py
This code simply indexes and displays specific classes in the CIFAR 100 Dataset. These classes are alphabetical and can be located on their website. Once a specific class is chosen the amount of images from that class can also be decided. Note that there is a random selection value this is used so that you can index a specific class multiple times and get different images if you would like to see the first images in that class set this feature to false.